move octant structs back to octant.go
change types file to deployment file

add some todos for the nastier error handling bits
